requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
open_clip_torch>=2.23.0
Pillow>=10.0.0
anthropic>=0.20.0
gradio>=4.0.0
matplotlib>=3.7.0
numpy>=1.24.0
requests>=2.31.0
huggingface_hub>=0.19.0
